fix: Add react router switch
This is to stop multiple routes being rendered.
Even though it was not showing the post detail component
was attempting to render. It did not because there was no post id
for the server to find. But when rendering the new post form the
server was being queried and throwing errors.